CRUD de cursos,añadir videos(Curso,actividad),CRUD de centros,mejora en la busqueda de actividad,CRUD de beneficios(añadir los no existentes)

+++ database/migrations/2026_04_10_150103_create_centros_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->string('imagen')->nullable();
             $table->text('descripcion');
             $table->boolean('es_activo')->default(true);
-            $table->float('latitud', 10, 8)->nullable();
-            $table->float('longitud', 11, 8)->nullable();
+            $table->decimal('latitud', 10, 8)->nullable();
+            $table->decimal('longitud', 11, 8)->nullable();
             $table->timestamps();
         });
     }
